requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
dnspython==2.7.0
fastapi==0.115.5
fonttools==4.55.0
h11==0.14.0
httptools==0.6.4
idna==3.10
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pydantic==2.10.1
pydantic_core==2.27.1
pymongo==4.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==0.24.0
websockets==14.1
